Update data in U*_P*_Imp*_SHA256Test::provideTestDetect()
The test-data-provider provideTestDetect() is missing to provide a
hash that contains characters in the salt out of the base64
range.

Previously only data with salts in the base64 range was provided.
This had left non-base64 ranged salts untested.

Steps done:

- A valid SHA256 hash containing characters in the salt out of the
  base64 range has been added.

For reference the following script shows the hash, what's its
plain is and that it verifies:

    <?php

    $plain  = 'hello';
    $hash   = '$5$' . "\xE4\"|\xF5|\x08\xC8'\xF054:>\x13\xCB\xED"
              . "\$6I5JMX./GN9KGHTtvHwvp3mxkNv/Ni7/jomOEBgsiM.";

    $verify = $hash === crypt($plain, $hash);
    if (!$verify) {
        throw new RuntimeException('PHP crypt() failure.');
    }

    echo "OK";

https://eval.in/38389