SHA256/SHA512 low-byte salts for tests
The test-data used in the imlementations:

 - Unit_Password_Implementation_SHA256Test
 - Unit_Password_Implementation_SHA512Test

did only contain salts inside the base64 range. That did
left non-base64 data for salts untested.

This patch adds hash strings that are using a salt
of 16 times chr(1) instead of chr(0). Both in the mock
of the random generator as well as in test-data.